break ecmdParseTokens() out into its own file, for testing purposes
