+++ ecmd-core/capi/ecmdParseTokens.C
@@ -1,0 +1,34 @@
+//IBM_PROLOG_BEGIN_TAG
+/* 
+ * Copyright 2003,2018 IBM International Business Machines Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * 	http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+//IBM_PROLOG_END_TAG
+#include <ecmdSharedUtils.H>
+
+void ecmdParseTokens (std::string line, const char* seperators, std::vector<std::string> & tokens) {
+
+  size_t curStart = 0, curEnd = 0;
+
+  tokens.clear();
+
+  while (1) {
+    curStart = line.find_first_not_of(seperators, curEnd);
+    if (curStart == std::string::npos) break;
+    curEnd = line.find_first_of(seperators,curStart);
+    tokens.push_back(line.substr(curStart, curEnd-curStart));
+  }
+}
+
